Updates the Dockerfile and install scripts
This update uses the versions as BUILD_ARGS so they can be overwritten
during the build step, but not persisted through to the container when
it's running.

I've also pulled the yum install and updates out of the install file and
added them to the Dockerfile so they can be properly layered.